ltx-video/backend/pyproject.toml
[project]
name = "ltx-video-backend"
version = "1.0.0"
description = "LTX-2 Video Generation Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "python-multipart>=0.0.9",
    "pillow>=10.3.0",
    "torch>=2.3.0",
    "huggingface-hub>=0.23.0",
    "tqdm>=4.66.0",
    "pynvml>=11.5.0",
    # LTX-2 inference dependencies
    "ltx-core",
    "ltx-pipelines",
    "diffusers>=0.32.2",
    "ftfy>=6.0.0",
    "imageio>=2.37.2",
    "imageio-ffmpeg>=0.6.0",
    "peft>=0.13.2",
    "protobuf>=3.20.0",
    "sentencepiece>=0.1.99",
]

[tool.uv.sources]
ltx-core = { git = "https://github.com/LightricksResearch/ltx-2-internal.git", subdirectory = "packages/ltx-core", rev = "d10477aec456c9c70103ee276318c316076c5ec2" }
ltx-pipelines = { git = "https://github.com/LightricksResearch/ltx-2-internal.git", subdirectory = "packages/ltx-pipelines", rev = "d10477aec456c9c70103ee276318c316076c5ec2" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
